fix budget calculation
